Remove warning Converter

+++ src/Converter.cpp
@@ -25,7 +25,7 @@ namespace Converter {
 
 std::vector<cv::Mat> toDescriptorVector(const cv::Mat &Descriptors) {
   std::vector<cv::Mat> vDesc;
-  vDesc.reserve(Descriptors.rows);
+  vDesc.reserve(static_cast<size_t>(Descriptors.rows));
   for(int j = 0; j < Descriptors.rows; j++) {
     vDesc.push_back(Descriptors.row(j));
   }
@@ -35,10 +35,12 @@ std::vector<cv::Mat> toDescriptorVector(const cv::Mat &Descriptors) {
 
 g2o::SE3Quat toSE3Quat(const cv::Mat &cvT) {
   Eigen::Matrix<double, 3, 3> R;
-  R << cvT.at<float>(0, 0), cvT.at<float>(0, 1), cvT.at<float>(0, 2), cvT.at<float>(1, 0), cvT.at<float>(1, 1),
-    cvT.at<float>(1, 2), cvT.at<float>(2, 0), cvT.at<float>(2, 1), cvT.at<float>(2, 2);
+  R << static_cast<double>(cvT.at<float>(0, 0)), static_cast<double>(cvT.at<float>(0, 1)), static_cast<double>(cvT.at<float>(0, 2)),
+    static_cast<double>(cvT.at<float>(1, 0)), static_cast<double>(cvT.at<float>(1, 1)), static_cast<double>(cvT.at<float>(1, 2)),
+    static_cast<double>(cvT.at<float>(2, 0)), static_cast<double>(cvT.at<float>(2, 1)), static_cast<double>(cvT.at<float>(2, 2));
 
-  const Eigen::Matrix<double, 3, 1> t(cvT.at<float>(0, 3), cvT.at<float>(1, 3), cvT.at<float>(2, 3));
+  const Eigen::Matrix<double, 3, 1> t(
+    static_cast<double>(cvT.at<float>(0, 3)), static_cast<double>(cvT.at<float>(1, 3)), static_cast<double>(cvT.at<float>(2, 3)));
 
   return g2o::SE3Quat(R, t);
 }
@@ -50,7 +52,7 @@ cv::Mat toCvMat(const g2o::SE3Quat &SE3) {
 
 cv::Mat toCvMat(const g2o::Sim3 &Sim3) {
   const Eigen::Matrix3d eigR = Sim3.rotation().toRotationMatrix();
-  const Eigen::Vector3d& eigt = Sim3.translation();
+  const Eigen::Vector3d &eigt = Sim3.translation();
   const double s = Sim3.scale();
   return toCvSE3(s * eigR, eigt);
 }
@@ -59,7 +61,7 @@ cv::Mat toCvMat(const Eigen::Matrix<double, 4, 4> &m) {
   cv::Mat cvMat(4, 4, CV_32F);
   for(int i = 0; i < 4; i++) {
     for(int j = 0; j < 4; j++) {
-      cvMat.at<float>(i, j) = m(i, j);
+      cvMat.at<float>(i, j) = static_cast<float>(m(i, j));
     }
   }
 
@@ -70,7 +72,7 @@ cv::Mat toCvMat(const Eigen::Matrix3d &m) {
   cv::Mat cvMat(3, 3, CV_32F);
   for(int i = 0; i < 3; i++) {
     for(int j = 0; j < 3; j++) {
-      cvMat.at<float>(i, j) = m(i, j);
+      cvMat.at<float>(i, j) = static_cast<float>(m(i, j));
     }
   }
 
@@ -80,7 +82,7 @@ cv::Mat toCvMat(const Eigen::Matrix3d &m) {
 cv::Mat toCvMat(const Eigen::Matrix<double, 3, 1> &m) {
   cv::Mat cvMat(3, 1, CV_32F);
   for(int i = 0; i < 3; i++) {
-    cvMat.at<float>(i) = m(i);
+    cvMat.at<float>(i) = static_cast<float>(m(i));
   }
 
   return cvMat.clone();
@@ -90,12 +92,12 @@ cv::Mat toCvSE3(const Eigen::Matrix<double, 3, 3> &R, const Eigen::Matrix<double
   cv::Mat cvMat = cv::Mat::eye(4, 4, CV_32F);
   for(int i = 0; i < 3; i++) {
     for(int j = 0; j < 3; j++) {
-      cvMat.at<float>(i, j) = R(i, j);
+      cvMat.at<float>(i, j) = static_cast<float>(R(i, j));
     }
   }
 
   for(int i = 0; i < 3; i++) {
-    cvMat.at<float>(i, 3) = t(i);
+    cvMat.at<float>(i, 3) = static_cast<float>(t(i));
   }
 
   return cvMat.clone();
@@ -103,14 +105,14 @@ cv::Mat toCvSE3(const Eigen::Matrix<double, 3, 3> &R, const Eigen::Matrix<double
 
 Eigen::Matrix<double, 3, 1> toVector3d(const cv::Mat &cvVector) {
   Eigen::Matrix<double, 3, 1> v;
-  v << cvVector.at<float>(0), cvVector.at<float>(1), cvVector.at<float>(2);
+  v << static_cast<double>(cvVector.at<float>(0)), static_cast<double>(cvVector.at<float>(1)), static_cast<double>(cvVector.at<float>(2));
 
   return v;
 }
 
 Eigen::Matrix<double, 3, 1> toVector3d(const cv::Point3f &cvPoint) {
   Eigen::Matrix<double, 3, 1> v;
-  v << cvPoint.x, cvPoint.y, cvPoint.z;
+  v << static_cast<double>(cvPoint.x), static_cast<double>(cvPoint.y), static_cast<double>(cvPoint.z);
 
   return v;
 }
@@ -118,8 +120,11 @@ Eigen::Matrix<double, 3, 1> toVector3d(const cv::Point3f &cvPoint) {
 Eigen::Matrix<double, 3, 3> toMatrix3d(const cv::Mat &cvMat3) {
   Eigen::Matrix<double, 3, 3> M;
 
-  M << cvMat3.at<float>(0, 0), cvMat3.at<float>(0, 1), cvMat3.at<float>(0, 2), cvMat3.at<float>(1, 0), cvMat3.at<float>(1, 1),
-    cvMat3.at<float>(1, 2), cvMat3.at<float>(2, 0), cvMat3.at<float>(2, 1), cvMat3.at<float>(2, 2);
+  M << static_cast<double>(cvMat3.at<float>(0, 0)), static_cast<double>(cvMat3.at<float>(0, 1)),
+    static_cast<double>(cvMat3.at<float>(0, 2)), static_cast<double>(cvMat3.at<float>(1, 0)),
+    static_cast<double>(cvMat3.at<float>(1, 1)), static_cast<double>(cvMat3.at<float>(1, 2)),
+    static_cast<double>(cvMat3.at<float>(2, 0)), static_cast<double>(cvMat3.at<float>(2, 1)),
+    static_cast<double>(cvMat3.at<float>(2, 2));
 
   return M;
 }
@@ -128,12 +133,8 @@ std::vector<float> toQuaternion(const cv::Mat &M) {
   const Eigen::Matrix<double, 3, 3> eigMat = toMatrix3d(M);
   const Eigen::Quaterniond q(eigMat);
 
-  return {static_cast<float>(q.x()),
-          static_cast<float>(q.y()),
-          static_cast<float>(q.z()),
-          static_cast<float>(q.w())};
+  return {static_cast<float>(q.x()), static_cast<float>(q.y()), static_cast<float>(q.z()), static_cast<float>(q.w())};
 }
 
 }  // namespace Converter
 }  // namespace ORB_SLAM2
-
